requirements.txt
google-cloud-bigquery==2.32.*
google-cloud-storage==2.0.*
pyyaml==6.0
pytest==6.2.*
pytest-cov==3.0.*
pytest-env==0.6.*
pylint==2.12.*
jinja2==3.0.*
